Refactored and improved mail sending
Before this commit, the build queue sent email on job failures. Now, the
job exit code is tracked and if it changes from the previous build, a
callback will be called with the current and previous job as parameters.

This callback currently implements sending emails when jobs fail or
recover.