Job env refactoring and implemented 'clean_env' setting.
A job setting 'clean_env' (default = true) now causes the environment to
be cleaned (i.e. not inherited from the parent process). Only the PATH
is inherited. If the job configuration sets it to 'false', the
environment from the parent process is inherited.

This commit also cleans up and refactors the environment creation for
jobs.